@@ -16,7 +16,7 @@ Iugu.USER_AGENT = {
   lang_version: process.version,
   platform: process.platform,
   publisher: 'iugu',
-  uname: null
+  uname: null,
 };
 
 Iugu.USER_AGENT_SERIALIZED = null;
@@ -36,14 +36,12 @@ var resources = {
 
   // The following rely on pre-filled customer IDs:
   CustomerPaymentMethods: require('./resources/CustomerPaymentMethods'),
-  
 };
 
 Iugu.IuguResource = require('./IuguResource');
 Iugu.resources = resources;
 
 function Iugu(key, version) {
-
   if (!(this instanceof Iugu)) {
     return new Iugu(key, version);
   }
@@ -55,7 +53,7 @@ function Iugu(key, version) {
     basePath: Iugu.DEFAULT_BASE_PATH,
     version: Iugu.DEFAULT_API_VERSION,
     timeout: Iugu.DEFAULT_TIMEOUT,
-    dev: false
+    dev: false,
   };
 
   this._prepResources();
@@ -64,76 +62,64 @@ function Iugu(key, version) {
 }
 
 Iugu.prototype = {
-
-  setHost: function(host, port, protocol) {
+  setHost: function (host, port, protocol) {
     this._setApiField('host', host);
     if (port) this.setPort(port);
     if (protocol) this.setProtocol(protocol);
   },
 
-  setProtocol: function(protocol) {
+  setProtocol: function (protocol) {
     this._setApiField('protocol', protocol.toLowerCase());
   },
 
-  setPort: function(port) {
+  setPort: function (port) {
     this._setApiField('port', port);
   },
 
-  setApiVersion: function(version) {
+  setApiVersion: function (version) {
     if (version) {
       this._setApiField('version', version);
     }
   },
 
-  setApiKey: function(key) {
+  setApiKey: function (key) {
     if (key) {
-      this._setApiField(
-        'auth',
-        'Basic ' + new Buffer(key + ':').toString('base64')
-      );
+      this._setApiField('auth', 'Basic ' + Buffer.from(key + ':').toString('base64'));
     }
   },
 
-  setTimeout: function(timeout) {
-    this._setApiField(
-      'timeout',
-      timeout == null ? Iugu.DEFAULT_TIMEOUT : timeout
-    );
+  setTimeout: function (timeout) {
+    this._setApiField('timeout', timeout == null ? Iugu.DEFAULT_TIMEOUT : timeout);
   },
 
-  _setApiField: function(key, value) {
+  _setApiField: function (key, value) {
     this._api[key] = value;
   },
 
-  getApiField: function(key) {
+  getApiField: function (key) {
     return this._api[key];
   },
 
-  getConstant: function(c) {
+  getConstant: function (c) {
     return Iugu[c];
   },
 
-  getClientUserAgent: function(cb) {
+  getClientUserAgent: function (cb) {
     if (Iugu.USER_AGENT_SERIALIZED) {
       return cb(Iugu.USER_AGENT_SERIALIZED);
     }
-    exec('uname -a', function(err, uname) {
+    exec('uname -a', function (err, uname) {
       Iugu.USER_AGENT.uname = uname || 'UNKNOWN';
       Iugu.USER_AGENT_SERIALIZED = JSON.stringify(Iugu.USER_AGENT);
       cb(Iugu.USER_AGENT_SERIALIZED);
     });
   },
 
-  _prepResources: function() {
-
+  _prepResources: function () {
     for (var name in resources) {
-      this[
-        name[0].toLowerCase() + name.substring(1)
-      ] = new resources[name](this);
+      this[name[0].toLowerCase() + name.substring(1)] = new resources[name](this);
     }
-
-  }
-
+  },
 };
 
 module.exports = Iugu;
