@@ -15,7 +15,7 @@ function _Error(raw) {
 _Error.prototype = Object.create(Error.prototype);
 
 _Error.prototype.type = 'GenericError';
-_Error.prototype.populate = function(type, message) {
+_Error.prototype.populate = function (type, message) {
   this.type = type;
   this.message = message;
 };
@@ -26,10 +26,9 @@ _Error.extend = utils.protoExtend;
  * Create subclass of internal Error class
  * (Specifically for errors returned from Iugu's REST API)
  */
-var IuguError = _Error.IuguError = _Error.extend({
+var IuguError = (_Error.IuguError = _Error.extend({
   type: 'IuguError',
-  populate: function(raw) {
-
+  populate: function (raw) {
     // Move from prototype def (so it appears in stringified obj)
     this.type = this.type;
 
@@ -39,14 +38,13 @@ var IuguError = _Error.IuguError = _Error.extend({
     this.message = raw.message;
     this.detail = raw.detail;
     this.raw = raw;
-
-  }
-});
+  },
+}));
 
 /**
  * Helper factory which takes raw iugu errors and outputs wrapping instances
  */
-IuguError.generate = function(rawIuguError) {
+IuguError.generate = function (rawIuguError) {
   switch (rawIuguError.type) {
     case 'card_error':
       return new _Error.IuguCardError(rawIuguError);
