@@ -4,32 +4,30 @@ var iuguMethod = require('./IuguMethod');
 var utils = require('./utils');
 
 module.exports = {
-
   create: iuguMethod({
-    method: 'POST'
+    method: 'POST',
   }),
 
   list: iuguMethod({
-    method: 'GET'
+    method: 'GET',
   }),
 
   retrieve: iuguMethod({
     method: 'GET',
     path: '/{id}',
-    urlParams: ['id']
+    urlParams: ['id'],
   }),
 
   update: iuguMethod({
     method: 'PUT',
     path: '{id}',
-    urlParams: ['id']
+    urlParams: ['id'],
   }),
 
   // Avoid 'delete' keyword in JS
   del: iuguMethod({
     method: 'DELETE',
     path: '{id}',
-    urlParams: ['id']
-  })
-
+    urlParams: ['id'],
+  }),
 };
