@@ -7,7 +7,7 @@ var utils = require('./utils');
  * Create an API method from the declared spec.
  *
  * @param [spec.method='GET'] Request Method (POST, GET, DELETE, PUT)
- * @param [spec.path=''] Path to be appended to the API BASE_PATH, joined with 
+ * @param [spec.path=''] Path to be appended to the API BASE_PATH, joined with
  *  the instance's path (e.g. "charges" or "customers")
  * @param [spec.required=[]] Array of required arguments in the order that they
  *  must be passed by the consumer of the API. Subsequent optional arguments are
@@ -15,18 +15,17 @@ var utils = require('./utils');
  *  (preceeding the also-optional callback argument
  */
 module.exports = function iuguMethod(spec) {
-
-  var commandPath = utils.makeURLInterpolator( spec.path || '' );
+  var commandPath = utils.makeURLInterpolator(spec.path || '');
   var requestMethod = (spec.method || 'GET').toUpperCase();
   var urlParams = spec.urlParams || [];
 
-  return function() {
-  
+  return function () {
     var self = this;
     var args = [].slice.call(arguments);
 
     var callback = typeof args[args.length - 1] == 'function' && args.pop();
-    var auth = args.length > urlParams.length && utils.isAuthKey(args[args.length - 1]) ? args.pop() : null;
+    var auth =
+      args.length > urlParams.length && utils.isAuthKey(args[args.length - 1]) ? args.pop() : null;
     var data = utils.isObject(args[args.length - 1]) ? args.pop() : {};
     var urlData = this.createUrlData();
 
@@ -48,19 +47,33 @@ module.exports = function iuguMethod(spec) {
 
     var requestPath = this.createFullPath(commandPath, urlData);
 
-    self._request(requestMethod, requestPath, data, auth, function(err, response) {
+    // For GET requests, move params to query string and clear body
+    if (requestMethod === 'GET' && data && typeof data === 'object') {
+      var qsData = {};
+      for (var k in data) {
+        if (Object.prototype.hasOwnProperty.call(data, k)) qsData[k] = data[k];
+      }
+      if (qsData.hasOwnProperty('count') && !qsData.hasOwnProperty('limit')) {
+        qsData.limit = qsData.count;
+        delete qsData.count;
+      }
+      var qs = utils.stringifyRequestData(qsData);
+      if (qs && qs.length > 0) {
+        requestPath = requestPath + (requestPath.indexOf('?') === -1 ? '?' : '&') + qs;
+      }
+      data = {};
+    }
+
+    self._request(requestMethod, requestPath, data, auth, function (err, response) {
       if (err) {
         deferred.reject(err);
       } else {
         deferred.resolve(
-          spec.transformResponseData ?
-            spec.transformResponseData(response) :
-            response
+          spec.transformResponseData ? spec.transformResponseData(response) : response
         );
       }
     });
 
     return deferred.promise;
-
   };
 };
