@@ -5,5 +5,5 @@ var iuguMethod = IuguResource.method;
 
 module.exports = IuguResource.extend({
   path: 'charge',
-  includeBasic: ['create']
+  includeBasic: ['create'],
 });
