@@ -15,5 +15,5 @@ var IuguResource = require('../IuguResource');
  */
 module.exports = IuguResource.extend({
   path: 'customers/{customerId}/payment_methods',
-  includeBasic: ['create', 'list', 'retrieve', 'update', 'del']
+  includeBasic: ['create', 'list', 'retrieve', 'update', 'del'],
 });
