@@ -5,10 +5,8 @@ var iuguMethod = IuguResource.method;
 
 module.exports = IuguResource.extend({
   path: 'customers',
-  includeBasic: [
-    'create', 'list', 'retrieve', 'update', 'del'
-  ],
-  
+  includeBasic: ['create', 'list', 'retrieve', 'update', 'del'],
+
   /**
    * Customer: PaymentMethods methods
    */
@@ -16,31 +14,30 @@ module.exports = IuguResource.extend({
   createPaymentMethod: iuguMethod({
     method: 'POST',
     path: '/{customerId}/payment_methods',
-    urlParams: ['customerId']
+    urlParams: ['customerId'],
   }),
 
   listPaymentMethod: iuguMethod({
     method: 'GET',
     path: '/{customerId}/payment_methods',
-    urlParams: ['customerId']
+    urlParams: ['customerId'],
   }),
 
   retrievePaymentMethod: iuguMethod({
     method: 'GET',
     path: '/{customerId}/payment_methods/{paymentMethodId}',
-    urlParams: ['customerId', 'paymentMethodId']
+    urlParams: ['customerId', 'paymentMethodId'],
   }),
 
   updatePaymentMethod: iuguMethod({
     method: 'POST',
     path: '/{customerId}/payment_methods/{paymentMethodId}',
-    urlParams: ['customerId', 'paymentMethodId']
+    urlParams: ['customerId', 'paymentMethodId'],
   }),
 
   deletePaymentMethod: iuguMethod({
     method: 'DELETE',
     path: '/{customerId}/payment_methods/{paymentMethodId}',
-    urlParams: ['customerId', 'paymentMethodId']
+    urlParams: ['customerId', 'paymentMethodId'],
   }),
-
 });
