@@ -8,6 +8,6 @@ module.exports = IuguResource.extend({
 
   create_account: iuguMethod({
     method: 'POST',
-    path: 'create_account'
-  })
+    path: 'create_account',
+  }),
 });
