@@ -2,5 +2,5 @@
 
 module.exports = require('../IuguResource').extend({
   path: 'payment_token',
-  includeBasic: ['create']
+  includeBasic: ['create'],
 });
