@@ -5,11 +5,10 @@ var iuguMethod = IuguResource.method;
 module.exports = IuguResource.extend({
   path: 'plans',
   includeBasic: ['create', 'list', 'retrieve', 'update', 'del'],
-  
+
   retrieveByIdentifier: iuguMethod({
     method: 'GET',
     path: 'identifier/{identifier}/',
-    urlParams: ['identifier']
-  })
+    urlParams: ['identifier'],
+  }),
 });
-
