@@ -4,10 +4,6 @@ var IuguResource = require('../IuguResource');
 var iuguMethod = IuguResource.method;
 
 module.exports = IuguResource.extend({
-
   path: 'transfers',
-  includeBasic: [
-    'create', 'list'
-  ],
+  includeBasic: ['create', 'list'],
 });
-
