@@ -5,15 +5,12 @@ var querystring = require('querystring');
 var hasOwn = {}.hasOwnProperty;
 var toString = {}.toString;
 
-var ARRAY_
-
-var utils = module.exports = {
-
+var utils = (module.exports = {
   isAuthKey: function (key) {
     return typeof key == 'string' && /^(?:[a-z]{2}_)?[A-z0-9]{32}$/.test(key);
   },
 
-  isObject: function(o) {
+  isObject: function (o) {
     return toString.call(o) === '[object Object]';
   },
 
@@ -21,8 +18,7 @@ var utils = module.exports = {
    * Stringifies an Object, accommodating a single-level of nested objects
    * (forming the conventional key "parent[child]=value")
    */
-  stringifyRequestData: function(data) {
-
+  stringifyRequestData: function (data) {
     var output = [];
 
     for (var i in data) {
@@ -39,23 +35,24 @@ var utils = module.exports = {
             output.push(encode(i) + '=' + encode(''));
           }
         } else if (Array.isArray(data[i])) {
-          
           for (var a = 0, l = data[i].length; a < l; ++a) {
-
-            if (utils.isObject(data[i][a])) {
+            var item = data[i][a];
+            if (utils.isObject(item)) {
               var hasProps = false;
-              for (var ii in data[i][a]) {
-                if (hasOwn.call(data[i][a], ii)) {
+              for (var ii in item) {
+                if (hasOwn.call(item, ii)) {
                   hasProps = true;
-                  output.push(encode(i + '[]' + '[' + ii + ']') + '=' + encode(data[i][a][ii]));
+                  output.push(encode(i + '[]' + '[' + ii + ']') + '=' + encode(item[ii]));
                 }
               }
               if (!hasProps) {
                 output.push(encode(i + '[]') + '=' + encode(''));
               }
+            } else {
+              // Primitive array items
+              output.push(encode(i + '[]') + '=' + encode(item));
             }
           }
-
         } else {
           output.push(encode(i) + '=' + encode(data[i]));
         }
@@ -76,33 +73,37 @@ var utils = module.exports = {
    *   var fn = makeURLInterpolator('some/url/{param1}/{param2}');
    *   fn({ param1: 123, param2: 456 }); // => 'some/url/123/456'
    */
-  makeURLInterpolator: (function() {
+  makeURLInterpolator: (function () {
     var rc = {
-      '\n': '\\n', '\"': '\\\"',
-      '\u2028': '\\u2028', '\u2029': '\\u2029'
+      '\n': '\\n',
+      '\"': '\\\"',
+      '\u2028': '\\u2028',
+      '\u2029': '\\u2029',
     };
     return function makeURLInterpolator(str) {
       return new Function(
         'o',
-        'return "' + (
+        'return "' +
           str
-          .replace(/["\n\r\u2028\u2029]/g, function($0) {
-            return rc[$0];
-          })
-          .replace(/\{([\s\S]+?)\}/g, '" + encodeURIComponent(o["$1"]) + "')
-        ) + '";'
+            .replace(/["\n\r\u2028\u2029]/g, function ($0) {
+              return rc[$0];
+            })
+            .replace(/\{([\s\S]+?)\}/g, '" + encodeURIComponent(o["$1"]) + "') +
+          '";'
       );
     };
-  }()),
+  })(),
 
   /**
    * Provide simple "Class" extension mechanism
    */
-  protoExtend: function(sub) {
+  protoExtend: function (sub) {
     var Super = this;
-    var Constructor = hasOwn.call(sub, 'constructor') ? sub.constructor : function() {
-      Super.apply(this, arguments);
-    };
+    var Constructor = hasOwn.call(sub, 'constructor')
+      ? sub.constructor
+      : function () {
+          Super.apply(this, arguments);
+        };
     Constructor.prototype = Object.create(Super.prototype);
     for (var i in sub) {
       if (hasOwn.call(sub, i)) {
@@ -115,6 +116,5 @@ var utils = module.exports = {
       }
     }
     return Constructor;
-  }
-
-};
+  },
+});
