@@ -3,11 +3,9 @@
 var iugu = require('../testUtils').getSpyableIugu();
 var expect = require('chai').expect;
 
-describe('Transfers Resource', function() {
-
-  describe('retrieve', function() {
-
-    it('Sends the correct request', function() {
+describe('Transfers Resource', function () {
+  describe('retrieve', function () {
+    it('Sends the correct request', function () {
       /*
       iugu.transfers.retrieve('transferId1');
       expect(iugu.LAST_REQUEST).to.deep.equal({
@@ -17,39 +15,31 @@ describe('Transfers Resource', function() {
       });
       */
     });
-
   });
 
-  describe('create', function() {
-
-    it('Sends the correct request', function() {
-
+  describe('create', function () {
+    it('Sends the correct request', function () {
       iugu.transfers.create({
-        amount: 200, currency: 'usd', recipient: {}
+        amount: 200,
+        currency: 'usd',
+        recipient: {},
       });
       expect(iugu.LAST_REQUEST).to.deep.equal({
         method: 'POST',
         url: '/v1/transfers',
-        data: { amount: 200, currency: 'usd', recipient: {} }
+        data: { amount: 200, currency: 'usd', recipient: {} },
       });
-
     });
-
   });
 
-  describe('list', function() {
-
-    it('Sends the correct request', function() {
-
+  describe('list', function () {
+    it('Sends the correct request', function () {
       iugu.transfers.list();
       expect(iugu.LAST_REQUEST).to.deep.equal({
         method: 'GET',
         url: '/v1/transfers',
-        data: {}
+        data: {},
       });
-
     });
-
   });
-
 });
